chore(cli): remove findup-sync from package dir and move to utils
remove findup-sync from package dir and move to utils
cli(utils): added path-util to be used after next release
added path-util to be used after next release

ISSUES CLOSED: #805
